Rename renderer to rendering
