+++ lib/component.js
@@ -11,7 +11,7 @@ import {
   scheduleRender,
   registerRenderer,
   unregisterRenderer
-} from './renderer.js';
+} from './rendering.js';
 
 setOnTagDirtied(scheduleRender);
