+++ lib/rendering.js
